pyproject.toml
[project]
name = "my-expenses"
version = "0.1.0"
description = "My expenses things"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django>=5.1.2",
    "uvicorn>=0.31.1",
    "python-dotenv>=1.0.1",
    "djangorestframework>=3.15.2",
    "django-filter>=24.3",
    "djangorestframework-simplejwt>=5.3.1",
    "pydantic>=2.9.2",
    "psycopg2-binary>=2.9.10",
    "django-cleanup>=9.0.0",
    "django-storages>=1.14.4",
    "boto3>=1.35.54",
    "django-silk>=5.3.1",
    "django-cors-headers>=4.6.0",
    "gunicorn>=23.0.0",
]

[dependency-groups]
dev = [
    "coverage>=7.6.2",
    "factory-boy>=3.3.1",
    "image>=1.5.33",
    "ipdb>=0.13.13",
    "pillow>=10.4.0",
    "pyright>=1.1.384",
    "pytest-django>=4.9.0",
    "pytest>=8.3.3",
    "ruff>=0.6.9",
    "tblib>=3.0.0",
]

[tool.ruff]
# Exclude a variety of commonly ignored directories.
exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".git-rewrite",
    ".hg",
    ".ipynb_checkpoints",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".pyenv",
    ".pytest_cache",
    ".pytype",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    ".vscode",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "site-packages",
    "venv",
]

line-length = 88
indent-width = 4

target-version = "py313"

[tool.ruff.lint]
select = ["E4", "E7", "E9", "F"]
ignore = ["F403"]

fixable = ["ALL"]
unfixable = []
dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"
docstring-code-format = true
docstring-code-line-length = "dynamic"


[tool.pyright]
reportIncompatibleVariableOverride = false
typeCheckingMode = "basic"
